refactor(reflect-cli): move methods into better filenames
